move common class MongoTestHelper to a tools module
